pyproject.toml
[project]
name = "toroidal-attention"
version = "0.1.0"
description = "3D Toroidal Attention Mechanism for Large Language Models"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "datasets>=2.12.0",
    "numpy>=1.26.0,<2.0.0", # 1.26+ required for Python 3.12 compatibility
    "tqdm>=4.65.0",
    "pyyaml>=6.0.0",        # Moved from requirements.txt
]

[project.optional-dependencies]
dev = [
    "pytest>=7.3.0",
    "pytest-cov>=4.1.0",
    "matplotlib>=3.7.0",
    "tensorboard>=2.13.0",
    "ruff>=0.1.0",         # Fast linter + formatter
]
training = [
    "accelerate>=0.20.0", # For distributed training
    "wandb>=0.15.0",      # Experiment tracking
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["toroidal_attention*", "scripts*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = [
    "E501", # Line too long (handled by formatter)
    "N806", # Uppercase variables (standard for ML: B=batch, N=seq_len, etc)
    "E741", # Ambiguous variable name (l is ok in specific contexts)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v --cov=toroidal_attention --cov-report=term-missing"
